Move the shannon measures into a subdirectory
