+++ src/CMakeLists.txt
@@ -12,7 +12,7 @@ set(${PROJECT_NAME}_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/predictive_info.c
     ${CMAKE_CURRENT_SOURCE_DIR}/relative_entropy.c
     ${CMAKE_CURRENT_SOURCE_DIR}/separable_info.c
-    ${CMAKE_CURRENT_SOURCE_DIR}/shannon.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/shannon/shannon.c
     ${CMAKE_CURRENT_SOURCE_DIR}/transfer_entropy.c
     ${CMAKE_CURRENT_SOURCE_DIR}/utilities/binning.c
     ${CMAKE_CURRENT_SOURCE_DIR}/utilities/black_boxing.c
