+++ src/shannon/univariate.c
